move types dir into src dir
