+++ package.json
@@ -7,9 +7,8 @@
     "lib": "lib"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "npx nodemon src/app.ts",
-    "server": "ts-node src/app.ts",
+    "dev": "npx nodemon src/app.ts",
+    "start": "ts-node src/app.ts",
     "build": "tsc"
   },
   "author": "",
