+++ src/types/query.d.ts
